refactor(ui): extract shared routine-exercise detail text; normalize exercise-row typography
- Extract verbatim-identical 63-line detail-text block from ExerciseRowInSuperset
  and ExerciseRowWithConnector into routineExerciseDetailText() (plain function —
  no Compose APIs used). Third copy (SupersetExerciseItem) deleted in Task 3.1.
- Typography normalization: ExerciseRowInSuperset name titleSmall→titleMedium;
  ExerciseRowWithConnector name Bold→Medium, detail bodyMedium→bodySmall.
- Fix lens-duplication-7: WeightStepper RoundedCornerShape(Spacing.*) →
  MaterialTheme.shapes.{medium,extraSmall}; fold inline letterSpacing into style.
- Remove unused RoundedCornerShape + ProgramMode imports from affected files.
