+++ shared/src/commonMain/kotlin/com/devil/phoenixproject/presentation/components/ExerciseRowWithConnector.kt
@@ -15,7 +15,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.DragHandle
 import androidx.compose.material.icons.filled.MoreVert
@@ -35,7 +34,6 @@ import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.devil.phoenixproject.domain.model.ProgramMode
 import com.devil.phoenixproject.domain.model.RoutineExercise
 import com.devil.phoenixproject.domain.model.WeightUnit
 import org.jetbrains.compose.resources.stringResource
@@ -160,75 +158,12 @@ fun ExerciseRowWithConnector(
                     Text(
                         exercise.exercise.name,
                         style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Bold,
+                        fontWeight = FontWeight.Medium,
                     )
-                    // Display format depends on whether this is a timed exercise
-                    // Bodyweight = no cable accessories (handles, bar, rope, etc.) in equipment list
-                    val isBodyweight = !exercise.exercise.hasCableAccessory
-
-                    val exerciseText = if (isBodyweight) {
-                        // Bodyweight exercise - always duration, never reps (no cables engaged)
-                        val duration = exercise.duration ?: 30
-                        "${exercise.sets} sets x ${duration}s"
-                    } else if (exercise.duration != null) {
-                        // Timed cable exercise - show duration AND weight/progression
-                        val isEchoMode = exercise.programMode == ProgramMode.Echo
-                        val weightText = if (isEchoMode) {
-                            "Adaptive"
-                        } else {
-                            val weight = kgToDisplay(exercise.weightPerCableKg, weightUnit)
-                            val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lbs"
-                            "${weight.toInt()} $unitLabel"
-                        }
-                        val progressionText = when {
-                            exercise.progressionKg > 0 -> {
-                                val progWeight = kgToDisplay(exercise.progressionKg, weightUnit)
-                                val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lb"
-                                " (+${progWeight}$unitLabel)"
-                            }
-
-                            exercise.progressionKg < 0 -> {
-                                val regWeight = kgToDisplay(-exercise.progressionKg, weightUnit)
-                                val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lb"
-                                " (-${regWeight}$unitLabel)"
-                            }
-
-                            else -> ""
-                        }
-                        "${exercise.sets} sets x ${exercise.duration}s @ $weightText$progressionText"
-                    } else {
-                        // Rep-based exercise with weight
-                        val isEchoMode = exercise.programMode == ProgramMode.Echo
-                        val weightText = if (isEchoMode) {
-                            "Adaptive"
-                        } else {
-                            val weight = kgToDisplay(exercise.weightPerCableKg, weightUnit)
-                            val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lbs"
-                            "${weight.toInt()} $unitLabel"
-                        }
-                        // Handle AMRAP vs fixed reps display
-                        val repsText = if (exercise.isAMRAP) "AMRAP" else "${exercise.reps} reps"
-                        // Build progression/regression suffix if configured
-                        val progressionText = when {
-                            exercise.progressionKg > 0 -> {
-                                val progWeight = kgToDisplay(exercise.progressionKg, weightUnit)
-                                val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lb"
-                                " (+${progWeight}$unitLabel/rep)"
-                            }
-
-                            exercise.progressionKg < 0 -> {
-                                val regWeight = kgToDisplay(-exercise.progressionKg, weightUnit)
-                                val unitLabel = if (weightUnit == WeightUnit.KG) "kg" else "lb"
-                                " (-${regWeight}$unitLabel/rep)"
-                            }
-
-                            else -> ""
-                        }
-                        "${exercise.sets} sets x $repsText @ $weightText$progressionText"
-                    }
+                    val exerciseText = routineExerciseDetailText(exercise, weightUnit, kgToDisplay)
                     Text(
                         exerciseText,
-                        style = MaterialTheme.typography.bodyMedium,
+                        style = MaterialTheme.typography.bodySmall,
                         color = MaterialTheme.colorScheme.onSurfaceVariant,
                     )
                 }
